switching to our customized awesome_notifications
