--- app/lib/common/notifications/desktop.dart
@@ -1,17 +1,0 @@
-// import 'package:quick_notify/quick_notify.dart';
-
-// Future<void> initializeNotifications() async {}
-
-// Future<void> requestNotificationsPermissions() async {
-//   // nothing to be done?
-//   var requestPermission = await QuickNotify.requestPermission();
-//   print('requestPermission $requestPermission');
-// }
-
-// void setupNotificationsListeners() {
-//   // mobile.setupNotificationsListeners();
-// }
-
-// Future<void> notify(String? text) async {
-//   await QuickNotify.notify(title: 'Acter', content: text);
-// }
