+++ app/pubspec.yaml
@@ -118,10 +118,10 @@ dependencies:
   riverpod_infinite_scroll: ^1.0.9
   infinite_scroll_pagination: ^3.2.0
   awesome_notifications:
-    # path: "../../awesome_notifications"
-    git:
-      url: https://github.com/acterglobal/awesome_notifications
-      ref: acter-main
+    path: "../../awesome_notifications"
+    # git:
+    #   url: https://github.com/acterglobal/awesome_notifications
+    #   ref: acter-main
 
 dev_dependencies:
   flutter_test:
